tools/pyproject.toml
[tool.poetry]
name = "azure-workspace-tools"
version = "0.1.0"
description = ""
authors = ["DSP Workspaces <dsp-workspaces@broadinstitute.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
azure-core = "^1.27.0"
azure-identity = "^1.14.0"
google-auth = "^2.19.1"

[tool.poetry.group.dev.dependencies]
black = {extras = ["d"], version = "^23.3.0"}
pdbpp = "^0.10.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
